Add files via upload
This code generates a random number between 1 and 1 million. The code prompts the user to guess the number, provides feedback on whether the guess is high or low, asks the user to modify the guess until the correct number is guessed, and then informs the user of the number of guesses made.